polish: extract slash commands, status bar elapsed timer, fix restore_terminal
- Extract handle_slash_command + SlashAction into tui_commands.rs (362 lines)
  tui_app.rs now 695 lines (was 1048)
- Status bar shows elapsed time during inference (⏳ 3s)
- Fix restore_terminal: remove hacky mem::replace with dummy terminal
- Clean up clippy warnings (dead_code, too_many_arguments)

File sizes:
  tui_app.rs:      695 lines
  tui_commands.rs:  362 lines
  tui_render.rs:    394 lines
  tui_output.rs:     60 lines
  status_bar.rs:     94 lines
