+++ koda-cli/src/tui_app.rs
@@ -18,9 +18,9 @@
 //! ```
 
 use crate::input;
-use crate::repl::{self, ReplAction};
+use crate::repl;
 use crate::sink::UiEvent;
-use crate::tui::{self, SelectOption};
+use crate::tui_commands::{self, SlashAction};
 use crate::tui_render::TuiRenderer;
 use crate::widgets::status_bar::StatusBar;
 
@@ -79,6 +79,7 @@ fn map_confirmation(c: crate::confirm::Confirmation) -> ApprovalDecision {
 
 // ── Viewport drawing ─────────────────────────────────────────
 
+#[allow(clippy::too_many_arguments)]
 fn draw_viewport(
     frame: &mut ratatui::Frame,
     textarea: &TextArea,
@@ -87,6 +88,7 @@ fn draw_viewport(
     context_pct: u32,
     state: TuiState,
     queue_len: usize,
+    elapsed_secs: u64,
 ) {
     let area = frame.area();
     let [input_row, status_row] =
@@ -115,6 +117,9 @@ fn draw_viewport(
     if queue_len > 0 {
         sb = sb.with_queue(queue_len);
     }
+    if elapsed_secs > 0 {
+        sb = sb.with_elapsed(elapsed_secs);
+    }
     frame.render_widget(sb, status_row);
 }
 
@@ -137,17 +142,6 @@ fn init_terminal() -> Result<Term> {
 
 fn restore_terminal(terminal: &mut Term) {
     let _ = terminal.clear();
-    drop(std::mem::replace(
-        terminal,
-        // Dummy — we just need to call clear() on the original
-        Terminal::with_options(
-            CrosstermBackend::new(std::io::stdout()),
-            TerminalOptions {
-                viewport: Viewport::Inline(0),
-            },
-        )
-        .unwrap(),
-    ));
     let _ = crossterm::terminal::disable_raw_mode();
     // Erase leftover viewport lines
     print!("\x1b[{}A\x1b[J", VIEWPORT_HEIGHT);
@@ -287,6 +281,7 @@ pub async fn run(
     let mut pending_command: Option<String> = None;
     let mut silent_compact_deferred = false;
     let mut should_quit = false;
+    let mut inference_start: Option<std::time::Instant> = None;
 
     // Crossterm event stream for async key capture
     let mut crossterm_events = EventStream::new();
@@ -304,6 +299,7 @@ pub async fn run(
             ctx,
             tui_state,
             input_queue.len(),
+            inference_start.map(|s| s.elapsed().as_secs()).unwrap_or(0),
         );
     })?;
 
@@ -339,7 +335,7 @@ pub async fn run(
                         print!("\x1b[{}A\x1b[J", VIEWPORT_HEIGHT);
                         let _ = std::io::Write::flush(&mut std::io::stdout());
 
-                        let action = handle_slash_command(
+                        let action = tui_commands::handle_slash_command(
                             &input,
                             &mut config,
                             &provider,
@@ -433,6 +429,7 @@ pub async fn run(
 
                         // Run the inference turn as a pinned future
                         tui_state = TuiState::Inferring;
+                        inference_start = Some(std::time::Instant::now());
 
                         {
                             let turn =
@@ -452,6 +449,7 @@ pub async fn run(
                                         ctx,
                                         tui_state,
                                         input_queue.len(),
+                                        inference_start.map(|s| s.elapsed().as_secs()).unwrap_or(0),
                                     );
                                 })?;
 
@@ -560,7 +558,7 @@ pub async fn run(
 
                         // Turn completed — cleanup
                         tui_state = TuiState::Idle;
-                        renderer.stop_spinner();
+                        inference_start = None;
                         crate::interrupt::reset();
                         session.cancel = tokio_util::sync::CancellationToken::new();
 
@@ -636,6 +634,7 @@ pub async fn run(
                 ctx,
                 tui_state,
                 input_queue.len(),
+                inference_start.map(|s| s.elapsed().as_secs()).unwrap_or(0),
             );
         })?;
 
@@ -696,353 +695,4 @@ pub async fn run(
     );
 
     Ok(())
-}
-
-// ── Slash command handler ────────────────────────────────────
-
-enum SlashAction {
-    Continue,
-    Quit,
-}
-
-#[allow(clippy::too_many_arguments)]
-async fn handle_slash_command(
-    input: &str,
-    config: &mut KodaConfig,
-    provider: &Arc<RwLock<Box<dyn LlmProvider>>>,
-    session: &mut KodaSession,
-    shared_mode: &approval::SharedMode,
-    renderer: &mut TuiRenderer,
-    project_root: &std::path::Path,
-    agent: &Arc<KodaAgent>,
-    pending_command: &mut Option<String>,
-) -> SlashAction {
-    match repl::handle_command(input, config, provider).await {
-        ReplAction::Quit => SlashAction::Quit,
-        ReplAction::SwitchModel(model) => {
-            config.model = model.clone();
-            config.model_settings.model = model.clone();
-            let mut s = koda_core::approval::Settings::load();
-            let _ = s.save_last_provider(
-                &config.provider_type.to_string(),
-                &config.base_url,
-                &config.model,
-            );
-            println!("  \x1b[32m\u{2713}\x1b[0m Model set to: \x1b[36m{model}\x1b[0m");
-            SlashAction::Continue
-        }
-        ReplAction::PickModel => {
-            let prov = provider.read().await;
-            match prov.list_models().await {
-                Ok(models) if models.is_empty() => {
-                    println!(
-                        "  \x1b[33mNo models available from {}\x1b[0m",
-                        prov.provider_name()
-                    );
-                }
-                Ok(models) => {
-                    drop(prov);
-                    let current_idx = models
-                        .iter()
-                        .position(|m| m.id == config.model)
-                        .unwrap_or(0);
-                    let options: Vec<SelectOption> = models
-                        .iter()
-                        .map(|m| {
-                            let desc = if m.id == config.model {
-                                "\u{25c0} current".to_string()
-                            } else {
-                                String::new()
-                            };
-                            SelectOption::new(&m.id, desc)
-                        })
-                        .collect();
-                    match tui::select("\u{1f43b} Select a model", &options, current_idx) {
-                        Ok(Some(idx)) => {
-                            config.model = models[idx].id.clone();
-                            config.model_settings.model = config.model.clone();
-                            let mut s = koda_core::approval::Settings::load();
-                            let _ = s.save_last_provider(
-                                &config.provider_type.to_string(),
-                                &config.base_url,
-                                &config.model,
-                            );
-                            println!(
-                                "  \x1b[32m\u{2713}\x1b[0m Model set to: \x1b[36m{}\x1b[0m",
-                                config.model
-                            );
-                        }
-                        Ok(None) => println!("  \x1b[90mCancelled.\x1b[0m"),
-                        Err(e) => println!("  \x1b[31mTUI error: {e}\x1b[0m"),
-                    }
-                }
-                Err(e) => println!("  \x1b[31mFailed to list models: {e}\x1b[0m"),
-            }
-            SlashAction::Continue
-        }
-        ReplAction::SetupProvider(ptype, base_url) => {
-            crate::commands::handle_setup_provider(config, provider, ptype, base_url).await;
-            SlashAction::Continue
-        }
-        ReplAction::PickProvider => {
-            crate::commands::handle_pick_provider(config, provider).await;
-            SlashAction::Continue
-        }
-        ReplAction::ShowHelp => {
-            let commands = [
-                ("/agent", "List available sub-agents"),
-                ("/compact", "Summarize conversation to reclaim context"),
-                ("/cost", "Show token usage for this session"),
-                ("/diff", "Show git diff / review / commit message"),
-                ("/expand", "Show full output of last tool call (/expand N)"),
-                ("/mcp", "MCP servers: status / add / remove / restart"),
-                ("/memory", "View/save project & global memory"),
-                ("/model", "Pick a model interactively"),
-                ("/provider", "Switch LLM provider"),
-                ("/sessions", "List/resume/delete sessions"),
-                ("/trust", "Set approval mode (always / auto / never)"),
-                ("/verbose", "Toggle full tool output (on/off)"),
-                ("/exit", "Quit the session"),
-            ];
-            let options: Vec<SelectOption> = commands
-                .iter()
-                .map(|(cmd, desc)| SelectOption::new(*cmd, *desc))
-                .collect();
-            if let Ok(Some(idx)) = tui::select("\u{1f43b} Commands", &options, 0) {
-                let (cmd, _) = commands[idx];
-                *pending_command = Some(cmd.to_string());
-            }
-            println!();
-            println!(
-                "  \x1b[90mTips: @file to attach context \u{00b7} Shift+Tab to cycle mode \u{00b7} Ctrl+C to cancel \u{00b7} Ctrl+D to exit\x1b[0m"
-            );
-            SlashAction::Continue
-        }
-        ReplAction::ShowCost => {
-            match session.db.session_token_usage(&session.id).await {
-                Ok(u) => {
-                    let total = u.prompt_tokens
-                        + u.completion_tokens
-                        + u.cache_read_tokens
-                        + u.cache_creation_tokens;
-                    println!();
-                    println!("  \x1b[1m\u{1f43b} Session Cost\x1b[0m");
-                    println!();
-                    println!("  Prompt tokens:     \x1b[36m{:>8}\x1b[0m", u.prompt_tokens);
-                    println!(
-                        "  Completion tokens: \x1b[36m{:>8}\x1b[0m",
-                        u.completion_tokens
-                    );
-                    if u.cache_read_tokens > 0 {
-                        println!(
-                            "  Cache read tokens: \x1b[32m{:>8}\x1b[0m",
-                            u.cache_read_tokens
-                        );
-                    }
-                    if u.cache_creation_tokens > 0 {
-                        println!(
-                            "  Cache write tokens:\x1b[33m{:>8}\x1b[0m",
-                            u.cache_creation_tokens
-                        );
-                    }
-                    if u.thinking_tokens > 0 {
-                        println!(
-                            "  Thinking tokens:   \x1b[35m{:>8}\x1b[0m",
-                            u.thinking_tokens
-                        );
-                    }
-                    println!("  Total tokens:      \x1b[1m{total:>8}\x1b[0m");
-                    println!("  API calls:         \x1b[90m{:>8}\x1b[0m", u.api_calls);
-                    println!();
-                    println!("  \x1b[90mModel: {}\x1b[0m", config.model);
-                    println!("  \x1b[90mProvider: {}\x1b[0m", config.provider_type);
-                }
-                Err(e) => println!("  \x1b[31mError: {e}\x1b[0m"),
-            }
-            SlashAction::Continue
-        }
-        ReplAction::ListSessions => {
-            match session.db.list_sessions(10, project_root).await {
-                Ok(sessions) if sessions.is_empty() => {
-                    println!("  \x1b[90mNo other sessions found.\x1b[0m");
-                }
-                Ok(sessions) => {
-                    let current_idx = sessions
-                        .iter()
-                        .position(|s| s.id == session.id)
-                        .unwrap_or(0);
-                    let options: Vec<SelectOption> = sessions
-                        .iter()
-                        .map(|s| {
-                            let desc = if s.id == session.id {
-                                format!(
-                                    "{}  {} msgs  {}k tokens  \u{25c0} current",
-                                    s.created_at,
-                                    s.message_count,
-                                    s.total_tokens / 1000
-                                )
-                            } else {
-                                format!(
-                                    "{}  {} msgs  {}k tokens",
-                                    s.created_at,
-                                    s.message_count,
-                                    s.total_tokens / 1000
-                                )
-                            };
-                            SelectOption::new(&s.id[..8], desc)
-                        })
-                        .collect();
-                    match tui::select("\u{1f43b} Sessions", &options, current_idx) {
-                        Ok(Some(idx)) => {
-                            let target = &sessions[idx];
-                            if target.id == session.id {
-                                println!("  \x1b[90mAlready in this session.\x1b[0m");
-                            } else {
-                                session.id = target.id.clone();
-                                println!(
-                                    "  \x1b[32m\u{2713}\x1b[0m Resumed session \x1b[36m{}\x1b[0m  \x1b[90m{}  {} msgs\x1b[0m",
-                                    &target.id[..8],
-                                    target.created_at,
-                                    target.message_count,
-                                );
-                            }
-                        }
-                        Ok(None) => println!("  \x1b[90mCancelled.\x1b[0m"),
-                        Err(e) => println!("  \x1b[31mTUI error: {e}\x1b[0m"),
-                    }
-                    println!("  \x1b[90mDelete: /sessions delete <id>\x1b[0m");
-                }
-                Err(e) => println!("  \x1b[31mError: {e}\x1b[0m"),
-            }
-            SlashAction::Continue
-        }
-        ReplAction::DeleteSession(ref id) => {
-            if id == &session.id {
-                println!("  \x1b[31mCannot delete the current session.\x1b[0m");
-            } else {
-                match session.db.list_sessions(100, project_root).await {
-                    Ok(sessions) => {
-                        let matches: Vec<_> =
-                            sessions.iter().filter(|s| s.id.starts_with(id)).collect();
-                        match matches.len() {
-                            0 => println!("  \x1b[31mNo session found matching '{id}'.\x1b[0m"),
-                            1 => {
-                                let full_id = &matches[0].id;
-                                match session.db.delete_session(full_id).await {
-                                    Ok(true) => println!(
-                                        "  \x1b[32m\u{2713}\x1b[0m Deleted session {}",
-                                        &full_id[..8]
-                                    ),
-                                    Ok(false) => {
-                                        println!("  \x1b[31mSession not found.\x1b[0m")
-                                    }
-                                    Err(e) => {
-                                        println!("  \x1b[31mError: {e}\x1b[0m")
-                                    }
-                                }
-                            }
-                            n => println!(
-                                "  \x1b[31mAmbiguous: '{id}' matches {n} sessions. Be more specific.\x1b[0m"
-                            ),
-                        }
-                    }
-                    Err(e) => println!("  \x1b[31mError: {e}\x1b[0m"),
-                }
-            }
-            SlashAction::Continue
-        }
-        ReplAction::ResumeSession(ref id) => {
-            if session.id.starts_with(id) {
-                println!("  \x1b[90mAlready in this session.\x1b[0m");
-            } else {
-                match session.db.list_sessions(100, project_root).await {
-                    Ok(sessions) => {
-                        let matches: Vec<_> =
-                            sessions.iter().filter(|s| s.id.starts_with(id)).collect();
-                        match matches.len() {
-                            0 => println!("  \x1b[31mNo session found matching '{id}'.\x1b[0m"),
-                            1 => {
-                                let target = &matches[0];
-                                session.id = target.id.clone();
-                                println!(
-                                    "  \x1b[32m\u{2713}\x1b[0m Resumed session \x1b[36m{}\x1b[0m  \x1b[90m{}  {} msgs\x1b[0m",
-                                    &target.id[..8],
-                                    target.created_at,
-                                    target.message_count,
-                                );
-                            }
-                            n => println!(
-                                "  \x1b[31mAmbiguous: '{id}' matches {n} sessions. Be more specific.\x1b[0m"
-                            ),
-                        }
-                    }
-                    Err(e) => println!("  \x1b[31mError: {e}\x1b[0m"),
-                }
-            }
-            SlashAction::Continue
-        }
-        ReplAction::InjectPrompt(prompt) => {
-            *pending_command = Some(prompt);
-            SlashAction::Continue
-        }
-        ReplAction::Compact => {
-            crate::commands::handle_compact(&session.db, &session.id, config, provider, false)
-                .await;
-            SlashAction::Continue
-        }
-        ReplAction::McpCommand(ref args) => {
-            crate::commands::handle_mcp_command(args, &agent.mcp_registry, project_root).await;
-            SlashAction::Continue
-        }
-        ReplAction::SetTrust(mode_name) => {
-            let new_mode = if let Some(ref name) = mode_name {
-                ApprovalMode::parse(name)
-            } else {
-                crate::commands::pick_trust_mode(approval::read_mode(shared_mode))
-            };
-            if let Some(m) = new_mode {
-                approval::set_mode(shared_mode, m);
-                println!(
-                    "  \x1b[32m\u{2713}\x1b[0m Trust: \x1b[1m{}\x1b[0m \u{2014} {}",
-                    m.label(),
-                    m.description()
-                );
-            } else if let Some(ref name) = mode_name {
-                println!(
-                    "  \x1b[31m\u{2717}\x1b[0m Unknown trust level '{}'. Use: plan, normal, yolo",
-                    name
-                );
-            }
-            SlashAction::Continue
-        }
-        ReplAction::Expand(n) => {
-            match renderer.tool_history.get(n) {
-                Some(record) => {
-                    crate::display::print_tool_output_full(record);
-                }
-                None => {
-                    let total = renderer.tool_history.len();
-                    if total == 0 {
-                        println!("  \x1b[90mNo tool outputs recorded yet.\x1b[0m");
-                    } else {
-                        println!(
-                            "  \x1b[33mNo tool output #{n}. Have {total} recorded (use /expand 1\u{2013}{total}).\x1b[0m"
-                        );
-                    }
-                }
-            }
-            SlashAction::Continue
-        }
-        ReplAction::Verbose(v) => {
-            renderer.verbose = match v {
-                Some(val) => val,
-                None => !renderer.verbose,
-            };
-            let state = if renderer.verbose { "on" } else { "off" };
-            println!("  \x1b[36mVerbose tool output: {state}\x1b[0m");
-            SlashAction::Continue
-        }
-        ReplAction::Handled => SlashAction::Continue,
-        ReplAction::NotACommand => SlashAction::Continue,
-    }
 }
